Fixed drawingboard page not working
was not working due the it not being a Droppable but using ItemCard component which is a Draggable, causing "Could not find required context" error. Fixed by making a separate but similar component called Notes for the DrawingBoard items. This is a temp fix, until  we figure out how to make DrawingBoard a droppable, which is probably not possible